refactor: 상세 페이지 전용 컴포넌트를 _components로 정리
- TableOfContents 컴포넌트를 articles/[slug]/_components/로 이동
- ArticleNavigation 컴포넌트를 articles/[slug]/_components/로 이동
- 페이지별 의존성을 명확히 하여 구조 개선
