--- src/components/TableOfContents/types.ts
@@ -1,5 +1,0 @@
-export interface TocHeading {
-  id: string;
-  text: string;
-  level: number;
-}
